pyproject.toml
[project]
name = "django-apm"
version = "0.1.0"
readme = "README.md"
description = "A fully featured APM for django using django"
authors = [
    {name = "Leandro de Souza", email = "leandrodesouzadev@gmail.com"},
]
dependencies = [
    "Django>=4.1.1",
    "requests>=2.28.1",
    "celery>=5.2.7",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/plathanus-tech/django-apm"
download_url = "https://github.com/plathanus-tech/django-apm/releases/tag/v0.1.0"

[project.optional-dependencies]


[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
makemessages = {env_file = "default.env", cmd = "django-admin makemessages"}

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.8.0",
]
